# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'My API'
  version: '1.0'
paths:
...
  /api/game/search:
    get:
      tags:
        - Game
      summary: 'Tìm kiếm (Theo tên thể loại và giá)'
      description: 'Tìm kiếm trò chơi theo từ khóa hoặc giá từ một mức giá cụ thể'
      parameters:
        -
          name: keyword
          in: query
          description: 'Từ khóa để tìm kiếm trò chơi'
          required: false
          schema:
            type: string
        -
          name: fromPrice
          in: query
          description: 'Giá tối thiểu để tìm kiếm trò chơi'
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Danh sách trò chơi được tìm thấy'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '404':
          description: 'Không tìm thấy trò chơi nào'
  '/api/game/detail/{id}':
    get:
      tags:
        - Game
      summary: 'Chi tiết trò chơi'
      description: 'Hiển thị thông tin chi tiết của một trò chơi'
      operationId: 'App\Application\Controllers\GameController::showGameByID'
      parameters:
        -
          name: id
          in: path
          description: 'ID của trò chơi'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Thông tin chi tiết của trò chơi'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: 'Không tìm thấy trò chơi'
  /api/game/favorate:
    get:
      tags:
        - Game
      summary: 'Được yêu thích nhất'
      description: 'Trả về top 5 các game được yêu thích'
      operationId: 'App\Application\Controllers\GameController::showFavorate'
      responses:
        '200':
          description: OK
        '422':
          description: 'Missing Data'
  /api/game:
    post:
      tags:
        - Game
      summary: 'Tạo trò chơi'
      description: 'Tạo mới một trò chơi'
      operationId: 'App\Application\Controllers\GameController::createGame'
      requestBody:
        description: 'Dữ liệu trò chơi cần tạo'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
      responses:
        '200':
          description: 'Trò chơi đã được tạo thành công'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Tạo thành công', type: string }
                  data: { $ref: '#/components/schemas/Game' }
                type: object
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  '/api/game/{id}':
    delete:
      tags:
        - Game
      summary: 'Xóa trò chơi'
      description: 'Xóa một trò chơi'
      operationId: 'App\Application\Controllers\GameController::deleteGame'
      parameters:
        -
          name: id
          in: path
          description: 'ID của trò chơi cần xóa'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Trò chơi đã được xóa thành công'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Thông báo thành công', type: string }
                type: object
        '404':
          description: 'Không tìm thấy trò chơi với ID đã cung cấp'
      security:
        -
          bearer: []
  /api/genre:
    get:
      tags:
        - Genre
      summary: 'Hiển thị toàn bộ thể loại'
      description: 'Hiển thị tất cả các thể loại trò chơi'
      operationId: 'App\Application\Controllers\GenreController::showAllGenre'
      responses:
        '200':
          description: 'Danh sách các thể loại trò chơi được tìm thấy'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        '404':
          description: 'Không tìm thấy bất kỳ thể loại trò chơi nào'
    post:
      tags:
        - Genre
      summary: 'Tạo mới thể loại trò chơi'
      description: 'Tạo mới một thể loại trò chơi'
      operationId: 'App\Application\Controllers\GenreController::createGenre'
      requestBody:
        description: 'Dữ liệu thể loại cần tạo'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  description: 'Tên của thể loại trò chơi'
                  type: string
              type: object
      responses:
        '200':
          description: 'Thể loại trò chơi đã được tạo thành công'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Tạo thành công', type: string }
                  data: { $ref: '#/components/schemas/Genre' }
                type: object
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  '/api/genre/detail/{id}':
    get:
      tags:
        - Genre
      summary: 'Hiển thị thể loại theo trò chơi'
      description: 'Hiển thị thể loại của trò chơi với ID được cung cấp'
      operationId: 'App\Application\Controllers\GenreController::showGenreOfGameID'
      parameters:
        -
          name: id
          in: path
          description: 'ID của trò chơi'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Thể loại của trò chơi được tìm thấy'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        '404':
          description: 'Không tìm thấy thể loại của trò chơi'
  '/api/genre/{id}':
    delete:
      tags:
        - Genre
      summary: 'Xóa thể loại'
      description: 'Xóa thể loại trò chơi'
      operationId: 'App\Application\Controllers\GenreController::deleteGenre'
      parameters:
        -
          name: id
          in: path
          description: 'ID của loại trò chơi cần xóa'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Thể loại đã được xóa thành công'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Thông báo thành công', type: string }
                type: object
        '404':
          description: 'Không tìm thấy thể loại với ID đã cung cấp'
      security:
        -
          bearer: []
...
components:
  schemas:
    Game:
      required:
        - id
        - name
        - description
        - price
        - image
        - publisher_id
        - like
        - status
      properties:
        id:
          description: 'ID của trò chơi'
          type: integer
          format: int64
        name:
          description: 'Tên của trò chơi'
          type: string
        description:
          description: 'Mô tả của trò chơi'
          type: string
        price:
          description: 'Giá của trò chơi'
          type: number
          format: double
        image:
          description: 'Đường dẫn hình ảnh của trò chơi'
          type: string
        publisher_id:
          description: 'ID của nhà xuất bản'
          type: integer
        like:
          description: 'Số lượt thích của trò chơi'
          type: integer
        status:
          description: 'Trạng thái của trò chơi'
          type: string
        created_at:
          description: 'Thời điểm tạo bản ghi'
          type: string
          format: date-time
        updated_at:
          description: 'Thời điểm cập nhật bản ghi'
          type: string
          format: date-time
      type: object
    GameRequest:
      required:
        - name
        - description
        - price
        - image
        - publisher_id
        - like
        - status
      properties:
        name:
          description: 'Tên của trò chơi'
          type: string
        description:
          description: 'Mô tả của trò chơi'
          type: string
        price:
          description: 'Giá của trò chơi'
          type: number
          format: double
        image:
          description: 'Đường dẫn hình ảnh của trò chơi'
          type: string
        publisher_id:
          description: 'ID của nhà xuất bản'
          type: integer
        like:
          description: 'Số lượt thích của trò chơi'
          type: integer
        status:
          description: 'Trạng thái của trò chơi'
          type: string
      type: object
    Genre:
      required:
        - name
      properties:
        name:
          description: 'Tên của thể loại game'
          type: string
      type: object
...
  securitySchemes:
    bearer:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header